+++ probe-rs/src/architecture/arm/traits.rs
@@ -123,6 +123,17 @@ pub trait RawDapAccess {
         Ok(())
     }
 
+    /// Configures the probe for JTAG use (specifying IR lengths of each DAP).
+    fn configure_jtag(&mut self) -> Result<(), DebugProbeError> {
+        Ok(())
+    }
+
+    /// Send a specific output sequence over JTAG.
+    ///
+    /// This can only be used for output, and should be used to generate
+    /// the initial reset sequence, for example.
+    fn jtag_sequence(&mut self, cycles: u8, tms: bool, tdi: u64) -> Result<(), DebugProbeError>;
+
     /// Send a specific output sequence over JTAG or SWD.
     ///
     /// This can only be used for output, and should be used to generate