+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `dap-server`: In addition to `Elf` format, this adds support for binary formats `Bin`, `Hex`, and `Idf` (#1656).
+- Added PAC55XX series targets (#1655)
+- Added support for JTAG commands via CMSIS-DAP protocol (#1462)
 - `core`: Added PAC55XX series targets (#1655)
 - `core`: Stack unwinding can now unwind beyond (optionally nested) exception handlers (#1665).
   - ARMv6-M: Report Exception / Fault description, and Unwind the registers and next frames.